Split models.py to models.py and tokenizers.py
