+++ swem/tokenizers.py
@@ -1,0 +1,13 @@
+from typing import List
+
+import MeCab
+
+
+def tokenize_en(text: str) -> List[str]:
+    text_processed = text.replace('.', ' .').replace(',', ' ,')
+    return text_processed.replace('?', ' ?').replace('!', ' !').split()
+
+
+def tokenize_ja(text: str, args: str = '-O wakati') -> List[str]:
+    tagger = MeCab.Tagger(args)
+    return tagger.parse(text).strip().split(' ')
