+++ tests/test_models.py
@@ -3,21 +3,7 @@ import unittest
 import numpy as np
 import pytest
 
-from swem import models
-
-
-def test_tokenize_ja():
-    """ A simple test for swem.tokenize """
-    tokens = models.tokenize_ja('私はバナナです。')
-    assert tokens == ['私', 'は', 'バナナ', 'です', '。']
-
-
-def test_tokenize_en():
-    """ test for swem.models.tokenize_en """
-    text = 'This, is an implementation of SWEM.'
-    tokens = models.tokenize_en(text)
-    expected = ['This', ',', 'is', 'an', 'implementation', 'of', 'SWEM', '.']
-    assert tokens == expected
+from swem import models, tokenizers
 
 
 def test_word_embed():
@@ -59,7 +45,7 @@ class SWEMTests(unittest.TestCase):
     def test_init_tokenizer_en(self):
         """ Passed lang='en', tokenizer equals to tokenize_en """
         model = models.SWEM(MockW2V(), lang='en')
-        assert model.tokenizer == models.tokenize_en
+        assert model.tokenizer == tokenizers.tokenize_en
 
     def test_infer_vector(self):
         methods = {
